refactor(user): move user constants to entities layer
- Move GENDER_OPTIONS, EDUCATION_OPTIONS, EXPERTISE_AREAS, INTEREST_AREAS, and REFERRAL_SOURCES from widgets/auth/config to entities/user/model
- Update import paths in onboarding steps to use new location
- Add proper exports in entities/user/model/index.ts
- Remove obsolete onboarding-options export from auth config

Improves code organization by placing user-related constants in the appropriate domain layer.
