+++ entities/user/index.ts
@@ -1,2 +1,3 @@
 export { getUserData, getUserDataKey } from './api/user-data';
 export type { UserIdentifierType } from './api/user-data';
+export * from './model';
