+++ entities/user/model/constants.ts
@@ -2,21 +2,51 @@ import type { OptionProps } from '@/shared/ui/option';
 import type { UserValidatorsCreateUserSchemaProfessionalFieldItem } from '@/generated/models';
 
 /**
+ * 性別選項
+ */
+export const GENDER_OPTIONS: OptionProps[] = [
+  { value: 'male', label: '男性' },
+  { value: 'female', label: '女性' },
+  { value: 'other', label: '其他' },
+];
+
+/**
+ * 教育階段選項
+ */
+export const EDUCATION_OPTIONS: OptionProps[] = [
+  { value: 'university', label: '大學' },
+  { value: 'high', label: '高中' },
+  { value: 'other', label: '其他' },
+];
+
+/**
  * 專業領域選項
  */
 export const EXPERTISE_AREAS: OptionProps<UserValidatorsCreateUserSchemaProfessionalFieldItem>[] =
   [
-    { value: 'information_and_communication_technologies_icts', label: '資訊與資訊通信技術(ICT)' },
+    {
+      value: 'information_and_communication_technologies_icts',
+      label: '資訊與資訊通信技術(ICT)',
+    },
     { value: 'business_administration_and_law', label: '商業、管理與法律' },
     { value: 'arts_and_humanities', label: '藝術、創意與設計' },
-    { value: 'natural_sciences_mathematics_and_statistics', label: '科學與研究' },
-    { value: 'engineering_manufacturing_and_construction', label: '工程與製造' },
+    {
+      value: 'natural_sciences_mathematics_and_statistics',
+      label: '科學與研究',
+    },
+    {
+      value: 'engineering_manufacturing_and_construction',
+      label: '工程與製造',
+    },
     { value: 'health_and_welfare', label: '健康與醫學' },
     { value: 'education', label: '教育與學習' },
     { value: 'social_sciences_journalism_and_information', label: '社會科學' },
     { value: 'language_skills_and_knowledge', label: '語言' },
     { value: 'services', label: '客戶服務與餐飲' },
-    { value: 'agriculture_forestry_fisheries_and_veterinary', label: '農業與環境科學' },
+    {
+      value: 'agriculture_forestry_fisheries_and_veterinary',
+      label: '農業與環境科學',
+    },
     { value: 'others', label: '其他' },
   ];
