+++ widgets/auth/config/index.ts
@@ -1,3 +1,2 @@
 export * from './onboarding-steps';
-export * from './onboarding-options';
