+++ widgets/auth/ui/steps/referral-source-step.tsx
@@ -3,9 +3,9 @@
 import { useFormContext } from 'react-hook-form';
 import { RadioGroupWithForm } from '@/shared/ui/radio-group';
 import { cn } from '@/shared/lib/cn';
+import { REFERRAL_SOURCES } from '@/entities/user';
 import type { RenderOptionFn } from '@/shared/ui/option';
 import type { OnboardingFormData } from '../../model';
-import { REFERRAL_SOURCES } from '../../config';
 
 const renderReferralOption: RenderOptionFn = ({
   Option,
